% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity2sims_ppos.R
\name{approximate_posterior}
\alias{approximate_posterior}
\title{approximate_posterior}
\usage{
approximate_posterior(mod, moddat)
}
\arguments{
\item{mod}{Stan model}

\item{moddat}{Model data}
}
\value{
A list with posterior mean and covariance
}
\description{
Use Laplace approximation to posterior
}
