% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{array_as_dt}
\alias{array_as_dt}
\title{array_as_dt}
\usage{
array_as_dt(arr)
}
\arguments{
\item{arr}{An array}
}
\value{
The array converted to a data.table=
}
\description{
Convert a trial array into a data.table
}
