% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity2sims_ppos.R
\name{assess_futility}
\alias{assess_futility}
\title{assess_futility}
\usage{
assess_futility(mod, moddat, n, p, n_max, n_sim = 1, B = 500)
}
\arguments{
\item{mod}{The assumed model}

\item{moddat}{The data required for the model (excluding outcomes which will be simulated)}

\item{n}{The observed sample size in each arm}

\item{p}{The probability of each outcome level in each arm}

\item{n_max}{The maximum sample size in each arm}

\item{n_sim}{The number of simulations to run}

\item{B}{The number of imputations for PPoS}
}
\value{
A list with two elements:
\itemize{
\item the first element is an array of dimension B x nrow(p) - 1 x n_sim which each
element a predicted posterior probability that beta is less than zero
\item the second element is a matrix of posterior probabilities
}
}
\description{
Assess futility rule at given sample size.
}
\details{
The observation model is calculated using dynamic HMC.
Posterior predictive draws are generated from this model.
The PPoS calculation uses Laplace approximations for the imputed data posteriors.
}
