% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity2sims_ppos.R
\name{calc_ppos}
\alias{calc_ppos}
\title{calc_ppos}
\usage{
calc_ppos(mod, dat, yimp, delta = 0, n_draws = 0)
}
\arguments{
\item{mod}{An rstan model}

\item{dat}{The current data for the model}

\item{yimp}{The imputed sets (impute_data) of data to be combined with dat}

\item{delta}{Reference effect size}

\item{n_draws}{Number of posterior draws to use in calculating PPoS}
}
\description{
Calculate posterior probability of success.
}
\details{
Calculates PPoS using Laplace approximation
}
