% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compile_cmdstanr_mod}
\alias{compile_cmdstanr_mod}
\title{compile_cmdstanr_mod
Compile provided cumulative logistic Stan model for rstan}
\usage{
compile_cmdstanr_mod()
}
\value{
A \code{CmdStanModel} object
}
\description{
compile_cmdstanr_mod
Compile provided cumulative logistic Stan model for rstan
}
