% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compile_rstan_approx_mod}
\alias{compile_rstan_approx_mod}
\title{compile_rstan_mod
Compile provided cumulative logistic Stan model using alternative parameterisation for rstan.}
\usage{
compile_rstan_approx_mod()
}
\value{
A \code{stanmodel} object
}
\description{
compile_rstan_mod
Compile provided cumulative logistic Stan model using alternative parameterisation for rstan.
}
