% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_data}
\alias{generate_data}
\title{generate_data}
\usage{
generate_data(alpha, beta, X)
}
\arguments{
\item{alpha}{Intercept parameters}

\item{beta}{Design matrix parameters}

\item{X}{Design matrix}
}
\value{
Simulated data y of length nrow(X)
}
\description{
Generate data where each participant is a single row.
}
