% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_data_agg}
\alias{generate_data_agg}
\title{generate_data_agg}
\usage{
generate_data_agg(n, alpha, eta)
}
\arguments{
\item{n}{Sample size}

\item{alpha}{Intercept parameters}

\item{eta}{Group specific shift on alpha}
}
\value{
Simulated data matrix y of dimension lenth(eta) x length(alpha) + y
}
\description{
Generate aggregated data where each row is counts across a covariate pattern.
}
