% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{generate_data_p}
\alias{generate_data_p}
\title{Generate ordinal data in aggregate}
\usage{
generate_data_p(n, p)
}
\arguments{
\item{n}{The sample size per arm}

\item{p}{The outcome level probabilities (columns) per arm (rows)}
}
\value{
A vector same dimension as p with outcome level counts by arm
}
\description{
Simulate data using p instead of alpha and eta/beta
}
