% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity2sims_ppos.R
\name{impute_data}
\alias{impute_data}
\title{impute_data}
\usage{
impute_data(npred, post_p)
}
\arguments{
\item{npred}{A vector of length N giving the number of future
participants assigned to treatment n}

\item{post_p}{An array of posterior draws of dimension N x K x B,
where N is treatment arms, K is outcome levels, and B is repetitions}
}
\value{
An array of same dimension as post_p, but with posterior-predicted
aggregated data.
}
\description{
Impute future data using posterior predictive
}
