% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_as_dt}
\alias{list_as_dt}
\title{list_as_dt}
\usage{
list_as_dt(res)
}
\arguments{
\item{res}{A list which is the result of simulate_longitudinal_trial}
}
\value{
The list converted to a data.table with one row being an analysis by treatment result
}
\description{
Convert a trial list into a trial data.table
}
