% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log1m}
\alias{log1m}
\title{log1m}
\usage{
log1m(x)
}
\arguments{
\item{x}{A real valued input}
}
\value{
The result of log(1 - x)
}
\description{
Calculate log(1 - x)
See https://cran.r-project.org/web/packages/Rmpfr/vignettes/log1mexp-note.pdf
}
