% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{log_diff_exp}
\alias{log_diff_exp}
\title{log_diff_exp}
\usage{
log_diff_exp(x, y)
}
\arguments{
\item{x}{A real valued input}

\item{y}{A real valued input}
}
\value{
The result of log(exp(x) - exp(y))
}
\description{
Calculate log(exp(x) - exp(y))
}
