% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ordinal_probs}
\alias{ordinal_probs}
\title{ordinal_probs}
\usage{
ordinal_probs(alpha)
}
\arguments{
\item{alpha}{A vector of cut-points (increasing)}
}
\value{
A vector of level probabilities
}
\description{
Calculate marginal probabilities from cut-points
}
