% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ordinal_probs_mats}
\alias{ordinal_probs_mats}
\title{ordinal_probs_mat}
\usage{
ordinal_probs_mats(alpha, eta)
}
\arguments{
\item{alpha}{A vector of cut-points (increasing)}

\item{eta}{A vector of linear predictors used to shift alpha (proportional odds)}
}
\value{
A vector of level probabilities
}
\description{
Calculate marginal probabilities matrix from cut-points
}
