% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity2sims.R
\name{sim_clarity2_trial}
\alias{sim_clarity2_trial}
\title{sim_clarity2_trial}
\usage{
sim_clarity2_trial(
  mod,
  n_seq = seq(600, 2100, 300),
  p_assign = rep(1/3, 3),
  alpha = stats::qlogis(cumsum(c(16, 29, 32, 13, 2, 1, 1, 6)/100)[1:7]),
  eta = c(0, -0.5, 0.5),
  eff_eps = 0.975,
  ...
)
}
\arguments{
\item{mod}{A list (model, model data)}

\item{n_seq}{Sequence of interim analysis sample sizes}

\item{p_assign}{Assignment probabilities to treatment arms}

\item{alpha}{True intercept parameter}

\item{eta}{True eta parameter}

\item{eff_eps}{Effectiveness threshold}

\item{...}{Other arguments to cmdstanr::sample, e.g. adapt_delta, chains, etc.}
}
\value{
A list of trial related data.tables
}
\description{
Simulate a clarity2 trial
}
\details{
Simulates a Clarity 2.0 trial.
}
