% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity2sims_ppos.R
\name{approximate_posterior}
\alias{approximate_posterior}
\title{approximate_posterior}
\usage{
approximate_posterior(mod, moddat, collapse_levels = TRUE)
}
\arguments{
\item{mod}{Stan model}

\item{moddat}{Model data}

\item{collapse_levels}{Optimisation likely to fail if any outcome levels have zero observations.
If this parameter is TRUE (default) then drops outcome levels with zero observations and shares
the prior weight equally across all remaining levels}
}
\value{
A list with posterior mean and covariance
}
\description{
Use Laplace approximation to posterior
}
